requirements.txt
vizdoom
boto3
numpy
pillow